rename source file for optimal inttypes
