+++ software/firmware/lib/include/stdint_optimized.h
